+++ PerfGate.js
@@ -27,7 +27,7 @@
  * MIT License
  */
 
-export const VERSION = '1.4.0';
+export const VERSION = '1.4.1';
 
 import {PerformanceObserver, constants} from 'node:perf_hooks';
 import {setTimeout as sleep} from 'node:timers/promises';